backend/requirements.txt
fastapi
uvicorn
aiofiles
requests
python-dotenv
openai

